+++ pygal/test/test_style.py
@@ -43,38 +43,3 @@ def test_parametric_styles_with_parameters():
     line.add('_', [1, 2, 3])
     line.x_labels = 'abc'
     assert line.render()
-
-
-def test_stroke_style():
-    s = Style(stroke_style='round')
-    assert s.stroke_style == 'round'
-    s = Style(stroke_style='bevel')
-    assert s.stroke_style == 'bevel'
-    s = Style(stroke_style='miter')
-    assert s.stroke_style == 'miter'
-    s = Style(stroke_style='rounded')
-    assert s.stroke_style == 'round'
-    s = Style(stroke_style='invalid derp')
-    assert s.stroke_style == 'round'
-
-
-def test_stroke_dasharray():
-    s = Style(stroke_dasharray=(0, 0))
-    assert s.stroke_dasharray == '0,0'
-    s = Style(stroke_dasharray=(.5, .5))
-    assert s.stroke_dasharray == '0,0'
-    s = Style(stroke_dasharray=(.9, .9))
-    assert s.stroke_dasharray == '0,0'
-    s = Style(stroke_dasharray=(1.9, 1.9))
-    assert s.stroke_dasharray == '1,1'
-
-
-def test_stroke_dasharray_input_types():
-    s = Style(stroke_dasharray=(0, 0))
-    assert s.stroke_dasharray == '0,0'
-    s = Style(stroke_dasharray='0,0')
-    assert s.stroke_dasharray == '0,0'
-    s = Style(stroke_dasharray='0x0')
-    assert s.stroke_dasharray == '0,0'
-    s = Style(stroke_dasharray='0  0')
-    assert s.stroke_dasharray == '0,0'